do not copy the WslNotifydWin unless the WslNotifyd restarts

+++ WslNotifydWin/scripts/runner.sh
@@ -12,6 +12,20 @@ shift
 windows_tmpdir_wsl="$(./scripts/get-windows-tmpdir.sh)"
 dst="${windows_tmpdir_wsl}/WslNotifydWin"
 
-rm -rf "${dst}"
-cp -r "${copy_src}" "${dst}"
+hash="$1"
+shift
+
+do_copy=1
+if [[ $hash != development && -f $dst/__hash__ ]]; then
+    hash_on_disk="$(<"${dst}/__hash__")"
+    if [[ $hash_on_disk = $hash ]]; then
+        do_copy=0
+    fi
+fi
+if [[ $do_copy = 1 ]]; then
+    echo "copy WslNotifyd to ${dst}"
+    rm -rf "${dst}"
+    cp -r "${copy_src}" "${dst}"
+fi
+
 exec "${dst}/WslNotifydWin.exe" "$@"
